requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.12.0
    # via
    #   -r requirements.in
    #   httpx
    #   starlette
cachetools==6.2.2
    # via
    #   -r requirements.in
    #   google-auth
certifi==2025.11.12
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
fastapi==0.115.4
    # via -r requirements.in
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   -r requirements.in
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via
    #   -r requirements.in
    #   google-generativeai
google-auth==2.45.0
    # via
    #   -r requirements.in
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-generativeai
    #   gspread
google-auth-httplib2==0.2.1
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.2
    # via gspread
google-generativeai==0.8.4
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gspread==6.2.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   -r requirements.in
    #   python-telegram-bot
idna==3.11
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   requests
oauthlib==3.3.1
    # via
    #   -r requirements.in
    #   requests-oauthlib
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.28.3
    # via
    #   -r requirements.in
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.10.3
    # via
    #   -r requirements.in
    #   fastapi
    #   google-generativeai
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.6.1
    # via -r requirements.in
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-telegram-bot==22.5
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
six==1.16.0
    # via -r requirements.in
starlette==0.41.3
    # via
    #   -r requirements.in
    #   fastapi
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
uvicorn==0.34.0
    # via -r requirements.in
